+++ eth2deposit/cli/existing_mnemonic.py
@@ -4,13 +4,13 @@ from typing import (
 )
 
 from eth2deposit.exceptions import ValidationError
-from eth2deposit.intl.utils import load_text
 from eth2deposit.key_handling.key_derivation.mnemonic import (
     verify_mnemonic,
 )
 from eth2deposit.utils.constants import (
     WORD_LISTS_PATH,
 )
+from eth2deposit.utils.intl import load_text
 from .generate_keys import (
     generate_keys,
     generate_keys_arguments_decorator,
