+++ eth2deposit/utils/constants.py
@@ -13,7 +13,7 @@ MAX_DEPOSIT_AMOUNT = 2 ** 5 * ETH2GWEI
 
 
 # File/folder constants
-INTL_CONETENT_PATH = os.path.join('eth2deposit', 'intl')
+INTL_CONTENT_PATH = os.path.join('eth2deposit', 'intl')
 WORD_LISTS_PATH = os.path.join('eth2deposit', 'key_handling', 'key_derivation', 'word_lists')
 DEFAULT_VALIDATOR_KEYS_FOLDER_NAME = 'validator_keys'
